Fix admin progress polling CI
